Add missing tests __init__ file
I'm tired of fighting this. Every tool under the sun wants the tests to
be an importable module, which doesn't make much sense to me, but it's
whatever. I'm just adding this because pylint, mypy, and others get
angry when I try to use them on just a directory without globbing